+++ golib.go
@@ -4,6 +4,8 @@ import (
 	"github.com/bigobject-inc/golib/irisweb"
 	"github.com/bigobject-inc/golib/logger"
 	"github.com/bigobject-inc/golib/miscellaneous"
+	"github.com/bigobject-inc/golib/space"
+	"github.com/bigobject-inc/golib/space/cctv"
 )
 
 // GoLibrary golang library service
@@ -14,4 +16,6 @@ type GoLibrary interface {
 	NewRotationLogger(loggerFile, loggerLevel string) (logger.Logger, error)
 	NewWeb(l logger.Logger, conf irisweb.Configure, rootPath string) (irisweb.Web, error)
 	NewWebController(path, method, name, summary string) (irisweb.Controller, error)
+	GetServGeometry() space.Geometry
+	NewCCTV(id string, geolocation [2]float64, resolution []interface{}, cameraMatrix [][]float64, rmat [][]float64, tvec [][]float64, dist [][]float64) cctv.CCTV
 }