+++ space/space.type.go
@@ -1,0 +1,15 @@
+package space
+
+//BoundingBox Geohash bounding box coordinates
+type BoundingBox struct {
+	NE     Point
+	SW     Point
+	Center Point
+}
+
+// Point World coordinates
+type Point struct {
+	Longitude float64 //x
+	Latitude  float64 //y
+	Height    float64 //z
+}