+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.0] - 2020-08-07
+
+### Added
+
+- Add space geometry
+- Add space cctv
+- Add GoLibrary.GetServGeometry
+- Add GoLibrary.NewCCTV
+
 ## [v1.0.3] - 2020-08-04
 
 ### Changed
@@ -35,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add logger
 - Add miscellaneous
 
+[v1.1.0]: https://github.com/bigobject-inc/golib/archive/v1.1.0.zip
 [v1.0.3]: https://github.com/bigobject-inc/golib/archive/v1.0.3.zip
 [v1.0.2]: https://github.com/bigobject-inc/golib/archive/v1.0.2.zip
 [v1.0.1]: https://github.com/bigobject-inc/golib/archive/v1.0.1.zip